修改报警

+++ pay-external-service/src/main/java/com/tfjt/pay/external/unionpay/dto/BusinessIsIncomingRespDTO.java
@@ -10,5 +10,5 @@ import lombok.Data;
 public class BusinessIsIncomingRespDTO {
     private String accountNo;
     private Long businessId;
-    private Integer businessType;
+    private Byte businessType;
 }
